// src/API.ts
export interface UsedBookOrderDto {
  buyerId?: number;
  buyerName?: string | null;
  id?: number;
  orderDate?: string;
  paymentMethod?: string | null;
  sellerId?: number;
  sellerName?: string | null;
  shippingFee?: number;
  status?: string | null;
  totalAmount?: number;
}

export interface UsedBookOrderDetailDto {
  bookID?: number;
  bookName?: string | null;
  id?: number;
  orderID?: number;
  unitPrice?: number;
}
// ...
export interface UsedBookBuyerInformation {
  id?: number;
  orderId: string;
  recipientAddress: string;
  recipientEmail?: string | null;
  recipientName: string;
  recipientPhone: string;
  remark?: string | null;
}